+++ rasa/utils/tf_models.py
@@ -1,3 +1,5 @@
+from collections import defaultdict
+
 import numpy as np
 import logging
 from typing import List, Text, Dict, Tuple, Union
@@ -6,7 +8,7 @@ from rasa.utils import train_utils
 from rasa.utils.common import is_logging_disabled
 import tensorflow as tf
 
-from rasa.utils.tf_model_data import RasaModelData
+from rasa.utils.tf_model_data import RasaModelData, FeatureSignature
 
 logger = logging.getLogger(__name__)
 
@@ -32,7 +34,6 @@ class RasaModel(tf.keras.models.Model):
         batch_size: Union[List[int], int],
         evaluate_on_num_examples: int,
         evaluate_every_num_epochs: int,
-        label_key: Text,
         batch_strategy: Text,
         silent: bool = False,
         eager: bool = False,
@@ -48,7 +49,7 @@ class RasaModel(tf.keras.models.Model):
             )
 
             model_data, evaluation_model_data = model_data.split(
-                evaluate_on_num_examples, random_seed, label_key=label_key
+                evaluate_on_num_examples, random_seed
             )
 
         disable = silent or is_logging_disabled()
@@ -57,12 +58,10 @@ class RasaModel(tf.keras.models.Model):
         tf_batch_size = tf.ones((), tf.int32)
 
         def train_dataset_function(x):
-            return model_data.as_tf_dataset(x, label_key, batch_strategy, shuffle=True)
+            return model_data.as_tf_dataset(x, batch_strategy, shuffle=True)
 
         def evaluation_dataset_function(x):
-            return evaluation_model_data.as_tf_dataset(
-                x, label_key, batch_strategy, shuffle=False
-            )
+            return evaluation_model_data.as_tf_dataset(x, batch_strategy, shuffle=False)
 
         if eager:
             # allows increasing batch size
@@ -193,3 +192,38 @@ class RasaModel(tf.keras.models.Model):
 
     def predict_generator(self, **kwargs) -> None:
         raise NotImplementedError
+
+    @staticmethod
+    def batch_to_model_data_format(
+        batch: Union[Tuple[np.ndarray], Tuple[tf.Tensor]],
+        data_signature: Dict[Text, List[FeatureSignature]],
+    ) -> Dict[Text, List[tf.Tensor]]:
+        """Convert input batch tensors into batch data format.
+    
+        Batch contains any number of batch data. The order is equal to the
+        key-value pairs in session data. As sparse data were converted into indices, data,
+        shape before, this methods converts them into sparse tensors. Dense data is
+        kept.
+        """
+
+        batch_data = defaultdict(list)
+
+        idx = 0
+        for k, signature in data_signature.items():
+            for is_sparse, shape in signature:
+                if is_sparse:
+                    # explicitly substitute last dimension in shape with known
+                    # static value
+                    batch_data[k].append(
+                        tf.SparseTensor(
+                            batch[idx],
+                            batch[idx + 1],
+                            [batch[idx + 2][0], batch[idx + 2][1], shape[-1]],
+                        )
+                    )
+                    idx += 3
+                else:
+                    batch_data[k].append(batch[idx])
+                    idx += 1
+
+        return batch_data
