+++ rasa/utils/train_utils.py
@@ -4,7 +4,7 @@ from typing import List, Optional, Text, Dict, Tuple, Union, Any, NamedTuple
 import numpy as np
 import tensorflow as tf
 
-from rasa.utils.tf_model_data import DataSignature
+from rasa.utils.tf_model_data import FeatureSignature
 
 logger = logging.getLogger(__name__)
 
@@ -22,40 +22,6 @@ def load_tf_config(config: Dict[Text, Any]) -> Optional[tf.compat.v1.ConfigProto
         return tf.compat.v1.ConfigProto(**config.pop("tf_config"))
     else:
         return None
-
-
-def batch_to_model_data_format(
-    batch: Union[Tuple[np.ndarray], Tuple[tf.Tensor]],
-    data_signature: Dict[Text, List[DataSignature]],
-) -> Dict[Text, List[tf.Tensor]]:
-    """Convert input batch tensors into batch data format.
-
-    Batch contains any number of batch data. The order is equal to the
-    key-value pairs in session data. As sparse data were converted into indices, data,
-    shape before, this methods converts them into sparse tensors. Dense data is
-    kept.
-    """
-
-    batch_data = defaultdict(list)
-
-    idx = 0
-    for k, signature in data_signature.items():
-        for is_sparse, shape in signature:
-            if is_sparse:
-                # explicitly substitute last dimension in shape with known static value
-                batch_data[k].append(
-                    tf.SparseTensor(
-                        batch[idx],
-                        batch[idx + 1],
-                        [batch[idx + 2][0], batch[idx + 2][1], shape[-1]],
-                    )
-                )
-                idx += 3
-            else:
-                batch_data[k].append(batch[idx])
-                idx += 1
-
-    return batch_data
 
 
 def confidence_from_sim(sim: "tf.Tensor", similarity_type: Text) -> "tf.Tensor":
